WIP Install extra dependencies for Swift
The list of dependencies for Swift is really small and it does not
include for example keystonemiddleware. This is not usually a problem
because keystonemiddleware is installed by something else when deploying
a complete OpenStack instance. It is a problem though when all services
but Swift are deployed using Python 3: in that case, the library
will not be available.
Swift recently added keystonemiddleware as "extra" dependency, so
use it (see https://review.openstack.org/544121).

Change-Id: Ifbbf2c561cbacf21a9f87f493e0e623f7c756e9a