WIP Add setup-networking role for multinode jobs
Multinode jobs need a mechanism for nodes to communicate to each
other so that they can support running a cloud with Neutron
networking.

Adding a setup-networking role that produces the same results as
the homonymous function in devstack-gate, and it includes most
of the code from the ovs-* roles in d-g.

Change-Id: I0f9381816ca775dd5ca0fb9153226f70201dd0aa