Add config to allow Glance not to run under uwsgi
Introduce a configuration option GLANCE_USE_UWSGI to control whether
Glance is deployed under uwsgi.  The option is False by default
because Glance's interoperable image import functionality and the
admin-only-by-default Tasks API do not work when Glance is run under
uwsgi, and it is desirable that Glance be deployed by default in
devstack in a fully working configuration.

Having the option available will allow easy experimentation for any
developers who wish to tackle implementing a fully working Glance
running in uwsgi.

Note: The Glance release notes for both the Pike and Queens releases
recommend that Glance not be run under uwsgi in production systems.

Change-Id: I8499d56901beed4f8288e3db813f871184388e49