Update patch set 1
Patch Set 1:

Not sure if the grenade failures are related to this change.  The default in devstack prior to this patch was glance running under uwsgi, now the default is glance running under eventlet.  Not sure why grenade tests would care.

Going to do a recheck because it's the weekend and there's not much load on the gates ATM.

Patch-set: 1