Fix multinode job
Multinode jobs require extra settings in devstack configuration,
some of which as subnode specific, some controller specific.
The job as setup by this patch matches the legacy multinode job
on devstack side (gate-tempest-dsvm-neutron-multinode-full-ubuntu-xenial-nv).

It also requires sharing the CA configuration between controller and
subnode, overlay network configuration for communication between
virtual machines and running discover_hosts for nova after the subnode
has been setup.

The extra orchestration required for multinode is encoded in a
dedicated role to allow for jobs in other repos to re-use it.

Depends-on: https://review.openstack.org/548026
Change-Id: I2dcbd9bdb401860820e655d97aa3c4775af2827f