Update patch set 1
Patch Set 1:

It would probably be good to link in the changelog for the libvirt 2.5.0 change with the spectre patch that is causing libvirt to take longer than normal to startup. That change seems to be unique to the ocata UCA which is what we're using in stable/pike and stable/ocata.

Patch-set: 1