Update patch set 1
Patch Set 1:

I'm not totally opposed to this, but by this point, devstack, tempest and whatever else has all finished, so this is a snapshot of whatever happens to be left-over; it's quite conceivable things have shutdown etc and the system is in a rather undefined state.

We have things like the peakmem tracker and worlddump.  What are you looking for?  Maybe we can find a way to enhance things in a more defined manner

Patch-set: 1
Label: Verified=0