WIP Install some more dependencies for Swift (python 2)
The list of dependencies for Swift is really small and it does not
include for example keystonemiddleware and ceilometermiddleware.
This is not usually a problem in a full-Python 2 deployment, but
it is when Swift is the only service where Python 3 is not enabled.

Add some workarounds to make sure that the missing libraries are
installed also in this case.

Change-Id: Ifbbf2c561cbacf21a9f87f493e0e623f7c756e9a