Update patch set 5
Patch Set 5: Code-Review-1

Is this something that needs to be fixed on opensuse side?
It seems odd to me not to use swif-init here to workaround a problem on a specific platform.
In any case I think we should have more comments in the code itself about the reason for this.

Patch-set: 5