[WIP][DNM] Add Ubuntu Xenial ARM64 platform job
This is intended for out-of-tree testing with zuul, and the extensive
modifications to drop jobs are unsuitable for deployment.  In the event
that this seems sensible as a platform job, the job should be defined
as an *addition* to other jobs, rather than as a replacement for so much.

In addition to the listed Depends-on, the job is not expected to succeed
without some manual activities to enable mirrors, etc.

Depends-on: https://review.openstack.org/546434
Change-Id: I6696b517068800668ccf99f95d43b1b64eee54ee