Update patch set 1
Patch Set 1: Code-Review-1

This actually can't ever land like this. As discussed during the PTG this causes a breaking change to configuration during an upgrade, and because this has been the status quo for 2 releases (and now rocky dev) it's not something we can just back away from. If you want to make this configurable it has to be as a separate new option that defaults to using uwsgi to preserve the upgrade path. This will enable users to opt-in to using eventlet.

As an aside this patch should be completely independent from proposed backports to fix branches, (since I've seen backports being abandoned because this was proposed) because changing the default is not something we can do easily. Especially since it's now the default on 3 branches. We want to provide a good experience for users of all projects (including devstack) so changing default behavior on a released branch is not something we can just. (even if CI allowed it, which here it is correctly blocking) However, backporting self contained fixes that comply to the stable policy to make the default experience work is the correct direction.

The quicker path moving forward is probably to just fix the async tasks path in glance under uwsgi. I've already started investigating this: https://review.openstack.org/#/c/549743/

But, since I've never used an api in glance that leverages an async task nor is there any tempest test coverage for those I don't have a mechanism to test which way works moving forward.

Patch-set: 1
Reviewer: Gerrit User 5196 <5196@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1