Update patch set 1
Patch Set 1:

If we're unable to convince the glance team that backporting a bug fix is a good idea (which still seems absurd that this is a point of contention) I think we should respin this to just inline the patch and automatically apply it on top of the glance tree. This is far from an ideal scenario, but without this fix and the corresponding glance patch pike devstack is not useable by a lot of people if they have images larger than cirros and/or are on a resource constrained system.

Patch-set: 1