Update patch set 1
Patch Set 1:

As I explained at the PTG there are multiple ways, but the primary one is the endpoint being used. Part of https://docs.openstack.org/grenade/latest/readme.html#theory-of-upgrade is that new code should run with old configs. Trying to change the endpoint between the old and new side of an upgrade won't work because everything is configured with the old one. 

But there are lots of other moving pieces that are all interconnected and just updating the daemon launched in the systemd unit file isn't going to mean that things seamlessly work exactly the same. The endpoint is just one example (and the most obvious one) but there are lots of places where things change between a uwsgi deployment and a standalone eventlet server.

This is why the proposed direction from our discussion at the PTG was on making this opt in and setting up a second job with the eventlet server so we can verify the proposed tempest test when that's ready. (Or simply with a depends-on patch series)

But that being said I think we can completely bypass the need for that extra work with a quick fix to the taskflow usage (as I started looking at with 549743) but I will need assistance verifying any proposed fix without a tempest test out there or some other script to test to exercise the api to launch an async task.

Patch-set: 1