Fail if RDO fails to install
It's not entirely clear if yum_install should be called outside of
"package_install"; at any rate it is here, and doesn't fail if the
package is missing, leading you on a weird chase trying to determine
why various packages aren't available during the run.  Do the simple
thing of exiting on failure.

Change-Id: I7cf3ed049306f43b2b22fb8f1a71ba0bb34638a9