Update patch set 2
Patch Set 2:

Yeah, we could manually deal with LIBS_FROM_GIT, but I feel like the user has already expressed what they want to have happen with required-projects, so I think we ought to be able to convince a computer to do that.

I've updated the patch to do the extra work to not add things we know will fail the check.

Patch-set: 2
Label: Verified=0