Update patch set 5
Patch Set 5:

I've uploaded PS5 which takes a different approach -- it makes the libs_from_git check safe for this method, at the expense of opening a slight opportunity for human error if a human writes a LIBS_FROM_GIT line.  I think the convenience is worth the risk (especially since the risk is reduced by using the new job and automated generation of the config).

Patch-set: 5
Label: Verified=0