Update patch set 2
Patch Set 2: Code-Review+1

I was also surprised to find this. The deprecation patch in d9aaae95f2b84170bf35e037715e4963d89f940c gave no background, but I assumed it was tied to the TC resolution.

There are jobs that rely on it so I don't see a reason to deprecate it, especially if it's not causing problems. We don't have a lot of postgresql bugs, no, but if one were reported, it's pretty easy to test the fix (it doesn't have to be on every single change, i.e. integrated gate).

Patch-set: 2
Label: Code-Review=+1