Update default ETCD_VERSION to latest 3.2 etcd release
This updates the default ETCD_VERSION to the latest 3.2 etcd release,
v3.2.17.  3.2 is chosen as it is packaged in bionic and fedora; we
hope to move to packaged versions for distros that support it in due
course.

This version supports arm64 and ppc64 which were not supported by the
previous default, v3.1.10.

We have removed the override to tarballs.o.o, as these files are now
cached as described in [1]

[1] http://lists.openstack.org/pipermail/openstack-infra/2018-March/005871.html

Change-Id: I5103b4331a8d7c5660848fc148ebe4139ce6dad9