WIP: Don't set database connection in controller in superconductor mode
CERN is reporting that nova-api won't function unless [database]/connection
is set in nova.conf which shouldn't really be necessary as the controller
services in nova (nova-api, nova-conductor (superconductor), nova-scheduler)
should rely on iterating cell mappings from the nova_api database.

This is to flush out any issues with that configuration.

Change-Id: I44811fde6e26271bb429bf4f1d59a9b094f66970
Related-Bug: #1757472