Update patch set 8
Patch Set 8: Code-Review+1

This looks good, thanks. We should probably check that this is working fine in https://review.openstack.org/#/c/548331/3.

Since this is a significant change in behavior compared to d-g, I think it would be good to publicize it a bit in the ML before it merges.

Patch-set: 8
Label: Code-Review=+1