Update patch set 8
Patch Set 8: Code-Review-1

(1 comment)

Also there is an inherent problem as to how the requirements are installed. When installing a package through this library it calls `_setup_package_with_constraints_edit` to make sure that this package is installed from source. However, it does not take into consideration that the {test-}requirements.txt might have packages that also might be present in LIBS_FROM_GIT. The end result is that those packages present in LIBS_FROM_GIT and in the requirements file will be installed them from the PyPi index instead.

I understand that this comment might exceed the scope of this patch, still it should be probably addressed because it has implications on this patch.

Patch-set: 8
Label: Code-Review=-1