Update patch set 12
Patch Set 12: Code-Review+2

so the issues is that SERVICE_HOST gets [] wrapped around it in stackrc, because in several places we do SERVICE_HOST:port and obviously that's an ipv6 issue

there seems to be some back-and-forth over if we should have []'s on by default, or add them when required, etc.  I share some of the same concerns over scattering the strippping function around v a global, but i see advantages and issues to both.  we have to make a choice, and this LGTM

Patch-set: 12
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2