Add option to install everything in global venvs
One venv for python2 and one for python3. This is a major change and
will likely break many things. But tl;dr is pip is making it
increasingly difficult to use globally install python packages with pip
alongside distro supplied packages. So add the option to install
everything into virtualenvs. This change was made optional in order to
add the required features to grenade without force merging updates;
however, we should consider making it non optional once we get the
basics in place.

Major hurdles we have to get over are convincing rootwrap to trust
binaries in the virtualenvs (so you'll notice we update rootwrap
configs). Additionally this doesn't yet support mixing python2 and
python3 installed services. You get python2 by default and python3 if
you opt into python3 but no mixing.

Depends-On: https://review.openstack.org/559418
Change-Id: If9bc7ba45522189d03f19b86cb681bb150ee2f25