Update patch set 8
Patch Set 8:

Michel, I believe the situation you describe shouldn't be a problem.  There are two cases:

* A libs-from-git project is installed via some project's requirements.txt file before the "if use_lib_from_git setup_develop" invocation.  In that case, the later "if use_lib_from_git setup_develop" call will reinstall the project from git.

* A libs-from-git project appears in some project's requirements.txt file after the "if use_lib_from_git setup_develop" invocation.  In that case, the previous setup_develop call would have updated the constraints file to write the git version, therefore the existing installation will match the constraint and nothing will change.

It's true that if anything goes wrong with that system, this change removes the final sanity check that would catch that, however, that's part of the trade-off I'm proposing we make.  And since both of those cases seem covered, we aren't relying on the final check to catch an expected error.

Patch-set: 8
Label: Verified=0