Update patch set 10
Patch Set 10:

The multinode job seems to be failing because tools/discover_hosts.sh is trying to call nova-manage without the proper path[0]. Not sure what is the best way to fix this, directly in the script, in the ds-gate wrapper or in the job definition.

I think the grenade jobs suffer from a similar issue, but haven't checked in detail yet.

[0] http://logs.openstack.org/30/558930/10/check/neutron-tempest-multinode-full/9ab9544/job-output.txt.gz#_2018-04-05_16_11_21_117757

Patch-set: 10
Label: Verified=0