Update patch set 2
Patch Set 2: Code-Review+2

Thanks for this.

I wonder if the REQUIREMENT_DIR is going to be the right one (master) on stable branches, but since this is the same behaviour we have for Tempest today (on stable/queens as well), it's fine.

Patch-set: 2
Reviewer: Gerrit User 1921 <1921@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2