Switch from python-psutil to pip install psutil
pip 10 will break devstack because devstack was installing python-psutil for
dstat. This patch switches psutil to be installed with pip.

Note: This is purely a guess at how to fix this, please feel free to take this
patch and run with it if you have a better way.

Add an environment variable to service files that gives the configured
python environment for the service.  Use this in the peakmem tracker
service to ensure we run under the right python.

Change-Id: Ib3cbb055f9402794b4c92397e69fcfedbef803cf