Update patch set 1
Patch Set 1:

I'm really not sure how useful this is.  We're only putting this in so that nodepool's element caching can run tools/image_list.sh during the build in some fashion.  if you're running stack.sh it will still blow up as unsupported and not do anything useful.

My preference would be just not to enter the devstack path at all when building gentoo images in nodepool, and leave devstack as is

Patch-set: 1