Update patch set 1
Patch Set 1:

Note after discussion, the situation is that there are several other places that use the cirros version that devstack is saving into /opt/files

The images we cache in devstack are really only supposed to be used by devstack to be uploaded into it's glance after the cloud is brought up.  When we bump the devstack version we're not thinking of other users who we'll break when they're pinned to an old version.

The namespacing of this is sub-optimal, maybe if it was in a "devsatck" subdirectory people might have realised what's going on.

We do not have a terrific generic caching story.  we can drop source-repository files in the cache-devstack element; and maybe that's enough.  we could extend this however, such that during the build we figure out what to cache by looking inside project repositories, so people don't have to propose project-config changes.

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0