Use oslo-config-generator for keystone config
The configue_keystone function was relying on a sample configuration
file being provided by keystone source. The sample file is usually
out-of-date and requires constant maintenance in order to be accurate.
Instead of relying on a sample in the source tree, we can just
generate a sample configuration file from the in-code configuration
module that keystone has.

Change-Id: I7a1f879e9e242a11e2c4663ec116e33da28db7f5