Handle in-tree tempest plugins
A recent patch[1] enabled upper-constraints checking for tempest plugins.
This patch causes failures for some scenarios with in-tree tempest plugins
because pip does not allow applying constraints to local paths. The resulting
error is "installation from path or url cannot be constrained to a version".

This patch handles local paths and URLs without checking upper constraints
but maintains the functionality for packages by name.

[1] https://review.openstack.org/560946
[2] http://logs.openstack.org/39/541039/13/check/octavia-v1-dsvm-scenario/ \
    c801e2b/job-output.txt.gz#_2018-04-17_22_09_54_936113

Change-Id: I445e218ff6a7d739e5f144833853726deaaa1a3a