Update patch set 9
Patch Set 9: Workflow+1 Code-Review+2

I agree with Jim, this does not change the current behaviour, it only adds a way to trigger the same behaviour as before, but automatically, from job properties.
When devstack is executed outside of a zuul CI job it will behave the same as today, except for that check that happens in a different place, but that does not alter the install process.

Could you please follow up either here or better in IRC or via the ML to express your concerns with the installation process?

Patch-set: 9
Reviewer: Gerrit User 1921 <1921@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2
Label: Workflow=+1