feature flag: Backwards incompatible policies
A new configuration group ``[policy_feature_enabled]`` has
been added to Patrole which will be responsible for collecting
the feature flags to be used for policies that were renamed
or removed in a backwards-incompatible
way. These feature flags are:

    * create_port_fixed_ips_ip_address_policy (Neutron)
    * update_port_fixed_ips_ip_address_policy (Neutron)
    * limits_extension_used_limits_policy (Cinder)
    * volume_extension_volume_actions_attach_policy (Cinder)
    * volume_extension_volume_actions_reserve_policy (Cinder)
    * volume_extension_volume_actions_unreserve_policy (Cinder)

These feature flags will be supported until Pike release cycle
is EOL.

The motivation behind these feature flags is [0] which adds
Pike/Queens gating to Patrole. However, in Queens, Neutron
and Cinder renamed or removed a few policies in a backwards-
incompatible way. These policies can be reviewed here: [1].

[0] I76c4a9b8737bf94f230ab141def652b054120f3b
[1] e.g. http://logs.openstack.org/51/547851/4/check/patrole-member-pike/139c534/job-output.txt.gz#_2018-03-22_21_46_08_392229

Change-Id: I00bdeff9474c54d38b6d6844a041b305bec01ad8
Needed-By: Ia0d9847908a8e723446c16465d68cd7f622c04cc