Update iniset to handle mysql-cluster conf files
mysql-cluster uses config files in 2 unique ways:

 - uses a section with a space in it, eg [ndb default]

 - uses an option with no value and requires no equal sign, eg

   [mysqld]
   ndbcluster

The first case causes problems when using regular expressions to find
the section. The string needs to have its space changed to \s when
searching. This was missed by the current unit tests which didn't
try to modify a section/option with spaces in the section name.

The second case presents a problem because calling iniset with a null
value will place an equal sign after the option. This works in most cases,
but will create an error in mysql-cluster if there is an equal sign after
the option for the ndbcluster option.

This updates iniset to deal with these issues. The default behavior will
not change.

Change-Id: I49ca3906df78c3cfbad4012bcdf777f701c69382
Closes-Bug: 1768282
Implements: blueprint mysql-cluster-support