Enable devstack to optionally install mysql-cluster
This change will allow a devstack deployment to optionally install
mysql-cluster.

To enable mysql-cluster, you will need to update devstack's local.conf
file. The changes needed in local.conf are:

  Change: DATABASE_TYPE=mysql-cluster
  Add: mysql-cluster to ENABLED_SERVICES
  Remove: mysql from ENABLED_SERVICES

There are two environmental variables that can be used to point to the
mysql-cluster repository and the version to be used. If they are not
specified, the default values are used.

  MYSQL_CLUSTER_REPO_URL (default value is http://repo.mysql.com/apt/ubuntu)

  MYSQL_CLUSTER_VERSION (default value is mysql-cluster-7.5)

Currently, mysql-cluster devstack deployment has only been tested on Ubuntu
16.04 on a single node.

Depends-On: I49ca3906df78c3cfbad4012bcdf777f701c69382
Depends-On: I597d6ab50fe117e725f1d8de9b359ac9fbb79eb3
Depends-On: Idf6e1f65ddb07e85ce864477aead976fd9a6af27
Depends-On: I06e08a7d0788606143cb9d5a500ca6bb36c907a0
Depends-On: Id717406bdb34bec20c57ae9ae0e50607738fb91d
Change-Id: I95525e63c073eb4fc8f3b1092d371af5a6e008cb
Implements: blueprint mysql-cluster-support