Use packaged uwsgi on Fedora and Ubuntu
Building uwsgi from source was a workaround that was introduced a long
time ago, it doesn't seem like it is needed anymore and will actually
fail for Ubuntu 20.04.

Also it doesn't match what will happen for most real-world
installations, so let's try to get back to using distro packages. We'll
still use the source install for RHEL/Ubuntu, it remains to be tested
whether we can get back to using distro versions there, too.

Depends-On: https://review.opendev.org/731159
Change-Id: I82f539bfa533349293dd5a8ce309c9cc0ffb0393