Don't install uwsgi for keystone
This is already unconditionally installed via install_apache_uwsgi in
stack.sh; we don't need to install it again in keystone.

Additionally, uwsgi is just a binary in the path; it works on some
platforms when $KEYSTONE_BIN_DIR is /usr/bin and uwsgi is in there,
but not generically.  Just call the uwsgi from path.

Change-Id: I40d84cbdf68cf6bb5cba143b6c0c126cdb8a84d4