Use devstack patrole lib to enable feature flags
This patchset adds patrole lib to devstack. This is
because devstack is a branched project, allowing for the
possibility of feature flags, which is not possible in
patrole devstack plugin, which is hosted in patrole repo,
which is branchless.

A Patrole-side change is made to reference the devstack
functionality for installation [0].

[0] I374aa6e825fb3117c4ffbf706a9825eb6c60b642

Change-Id: I4b17889e433bc9c3303596c715632b54012aeb02
Needed-By: I374aa6e825fb3117c4ffbf706a9825eb6c60b642