Use devstack patrole lib to enable feature flags
This patchset adds patrole lib to devstack. This is
because devstack is a branched project, allowing for the
possibility of feature flags, which is not possible in
patrole devstack plugin, which is hosted in patrole repo,
which is branchless.

Patrole is an RBAC testing validation framework, that,
like Tempest, is branchless. And there is a need [0]
within Patrole to run against n-1 and n-2 stable
branches. But to do so, feature flags [1] are needed.
Feature flags, in turn, have a dependency around
using a devstack, which is branched, to toggle on or
off feature flags for the supported stable branches.

Tempest is not the appropriate place [2] to include
the feature flags as Patrole is its own dedicated
project that should handle its own feature flags
separately.

Thus, the following needs to happen to enable Patrole
n-1/n-2 gating:

1) Add Patrole installation functon to devstack (this PS)
2) Update Patrole devstack plugin installation to reference
   function introduced by this PS [3]
3) Add Patrole feature flags as needed [1]
4) Toggle feature flags as needed for stable branches [4]
5) Enable gating of stable branches [0]

A Patrole-side change is made to reference the devstack
functionality for installation [3].

[0] I76c4a9b8737bf94f230ab141def652b054120f3b
[1] Ia0d9847908a8e723446c16465d68cd7f622c04cc
[2] https://review.openstack.org/#/c/556709/4/doc/source/configuration.rst@72
[3] I374aa6e825fb3117c4ffbf706a9825eb6c60b642
[4] I00bdeff9474c54d38b6d6844a041b305bec01ad8

Change-Id: I4b17889e433bc9c3303596c715632b54012aeb02
Needed-By: I374aa6e825fb3117c4ffbf706a9825eb6c60b642