Update patch set 3
Patch Set 3: Code-Review-1

So I hate to be negative nancy about this ... but it's used in quite a few places [1].  I don't think we should commit until we've come up with a plan. 

It's all copy paste -- I don't know how much we really care or if these are really used with USE_VENV; but it is out there.  

Two thoughts are we could just fix it up in those projects, or we could replace it with a stub that gives a sensible error

[1] http://codesearch.openstack.org/?q=venv.*python_version&i=nope&files=&repos=

Patch-set: 3
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1