Update patch set 1
Patch Set 1:

I would say that they may fail, but not always:
https://review.openstack.org/#/c/570559/
https://review.openstack.org/#/c/568317/
https://review.openstack.org/#/c/550108/

About this review, the failures in the fedora and tumbleweed jobs are unrelated (missing yum, and an error computing default_route_dev in neutron-legecy respectively). centos-7 fails with an error in a volume test ('Volume 369e7ede-eab0-4b7c-a22b-3f3c0b54586c failed to restore and is in ERROR status') which failed in other jobs too (for example https://review.openstack.org/#/c/568317/).

tl;dr the failures of non-xenial jobs are unrelated to this change (also, opensuse-423 worked).

Patch-set: 1