Update patch set 2
Patch Set 2:

when installed in a virtualenv pip takes care of inserting the correct shebang for django-admin(.py).

If it is not a virtualenv then pip just writes a shebang delegating env to evaluate the correct python.

manage.py and django-admin are both fine here, since they end up calling the same underlying command. I used django-admin as it is used for other commands too.

Patch-set: 2