Use packaged uwsgi for Ubuntu
It's not *exactly* clear what the problem is with the packaged Xenial
version that led to packaged install with
I00fb1759e6988c7df0ce0f3df5ff1ce9fd7cd381.

The change says "Ubuntu xenial is back level on uwsgi so the proxy
doesn't actually work.".  However, I noticed that ubuntu's
2.0.12-5ubuntu3.1 package has picked up a patch [1]:

 * Re-add patch cherry-picked upstream to fix apache 2.4 integration
   with unix domain sockets (accidentally dropped in 2.0.12-1).

I suspect we may be able to use the packaged uwsgi now, and not have
to worry about building from source.

To use packaged versions, we need to switch the python plugin type to
uwsgi manually.  A variable is added for this.

[1] http://launchpadlibrarian.net/335082909/uwsgi_2.0.12-5ubuntu3.1_source.changes

Change-Id: I82f539bfa533349293dd5a8ce309c9cc0ffb0393