Update patch set 19
Patch Set 19:

Note in general; I'm not super sure about this patch.  I guess it comes down to "what are we testing"?

Compiling uwsgi from source has fixed problems (the original issue) and caused problems (it not being up-to-date enough for fedora which had local patches).  it's not clear to me either is particularly any better.

my slight feeling is that if we're not testing with the packaged uwsgi for platforms, then what hope do people have of deploying on these platforms we're testing on?  it seems like it would be better to know about problems; like "uwsgi as packaged for Xenial doesn't seem to work on s390x" rather than find out when someone tries to do this in a more production friendly matter.  but i would understand if reviewers don't think this way too. 

@andreas - if you want to put a skip around using the packaged sources for s390x with a comment "this doesn't work and we don't know why" (and that makes it work :) i'm fine with that too -- assuming other reviewers are onboard with using the packaged versions at all.

Patch-set: 19
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0