Update patch set 1
Patch Set 1:

Maybe adding some remark to https://docs.openstack.org/devstack/latest/guides/multinode-lab.html would help, though the config samples provided there should not suffer from this issue.

The default set of services that are enabled is meant to work for an all-in-one deployment, which is the most common use case. Multi-node deployments always need special config and I think there are too many possible variants for this to be automated in a sensible way.

Patch-set: 1