Update patch set 1
Patch Set 1: Code-Review-1

i agree this is dead and i'd be very surprised if anyone wants to maintain it.

but can we depend-on a change to remove it more completely?  for example it's still referenced in devsatck-gate (which is also on the way out, but anyway) [1] and still referenced in documentation about that [2].  codesearch might show other places where it pops up

[1] http://git.openstack.org/cgit/openstack-infra/devstack-gate/tree/devstack-vm-gate.sh#n819
[2] http://git.openstack.org/cgit/openstack-infra/system-config/tree/doc/source/devstack-gate.rst

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1